+++ src/components/providers/ClientProviders.tsx
@@ -13,9 +13,7 @@ export function ClientProviders({ children }: ClientProvidersProps) {
   return (
     <SessionProvider>
       <QueryProvider>
-        <UserProfileProvider>
-          {children}
-        </UserProfileProvider>
+        <UserProfileProvider>{children}</UserProfileProvider>
       </QueryProvider>
     </SessionProvider>
   )
