+++ src/hooks/use-mobile.ts
@@ -1,11 +1,15 @@
+'use client'
+
 import * as React from 'react'
 
 const MOBILE_BREAKPOINT = 768
 
 export function useIsMobile() {
   const [isMobile, setIsMobile] = React.useState<boolean | undefined>(undefined)
+  const [hasMounted, setHasMounted] = React.useState(false)
 
   React.useEffect(() => {
+    setHasMounted(true)
     const mql = window.matchMedia(`(max-width: ${MOBILE_BREAKPOINT - 1}px)`)
     const onChange = () => {
       setIsMobile(window.innerWidth < MOBILE_BREAKPOINT)
@@ -14,6 +18,10 @@ export function useIsMobile() {
     setIsMobile(window.innerWidth < MOBILE_BREAKPOINT)
     return () => mql.removeEventListener('change', onChange)
   }, [])
+
+  if (!hasMounted) {
+    return false
+  }
 
   return !!isMobile
 }
